Convert remaining galloc calls to an inline function xmalloc1.
The main advantage here is that the name is a reminder of what group
of functions it supplements. Also it reminds us that the purpose is to
ensure that at least one byte is allocated. The problem is that
xmalloc might return NULL if its argument equals zero. (I imagine this
is because some malloc implementations already did that when the
library standard was devised.)