Bug in PostScript Type 3 output with bitmap glyphs
I started using FontForge a couple days ago to create some example PostScript
Type 3 fonts using bitmap glyphs, and found a couple problems in the output.

First, setcachedevice is being called like this for bitmap glyphs:

  /H { 720 0 0 0 0 0 setcachedevice

The syntax of the setcachedevice operator is:

  Wx Wy Llx Lly Urx Ury setcachedevice

The bounding box described by Llx Lly Urx Ury is typically used by a PostScript
interpreter to determine the size of the bitmap to use to render and cache the
glyph.  Supplying four zeros will either cause the glyph to be uncacheable
(since these are uncolored glyphs, they are cacheable) or worse, cause a
PostScript interpreter to fail to render the glyph properly at all.  The
information is available to supply the proper bounding box.

The second problem I found is in the matrix supplied to the imagemask operator,
specifically the ty value.

70 710 translate 565 715 scale 113 143 true [113 0 0 -143 0 143] datasrc imagemask

The CharProcs are positioning the origin at the top left corner of the glyph
bounding box, and then scanning the image downwards.  In this case, it is not
necessary to translate the origin by ty.  That is only needed when the origin
is positioned in the lower right corner and the image is scanned top to bottom.
So the last number should be 0 and not the image height.