Some use of my new anchor point ordering in the C internals. There is commentary as part of the change, expressing my ignorance of what I am doing and suggesting other input. To this end, for reference I merely commented out, rather than deleted, some replaced code.
Also I am sure there are more places where re-ordering would be a good
thing to do. After an affine transformation is a good example, if it
is not already being done then. But, because setting the anchor points
was not done through a subroutine, it is tedious to track all these
instances down. Argh.