Simpler fix
Creating a font view when loading a font is too much hassle (ViewFont()
will create a new one, rightly, and telling it to not create new one
only when opening a UFO font is too much for useless stuff like UFO), so
just fix the specific cause of this crash and hope it does not crash
elsewhere for the same reason.

One day we should drop all this UFO crap and move it to an extension of
some sort.