Fix parsing format 2 pair positioning
When parsing format 2 pair positioning (AKA class kerning), limit first
glyphs to the one present in the coverage table.

Previously we were ignoring the coverage table and just using every
glyph in the first class definition table, however if it were empty
getClassDefTable() will return class 0 for each glyph (default) so we
end up applying the kerning to each glyph in the font, which is wrong.

Outputting empty first class definition table seems dubious to me, but
AFDKO seems to be doing that and HarfBuzz parses that just fine, so we
now parse it correctly as well.

However this code needs an overhaul as we are expanding the classes and
storing them as individual kerning pairs, which is not ideal.