Avoid crash in Glyph Info dialog
When changing glyph name we will check if there is a glyph with the same
name and show the user a dialog, if he cancels the Glyph Info dialog is
kept open, but the next time the user presses OK we will crash.

The crash happens in CI_CheckMetaData() because oldsc->name in the 2nd
time will be NULL, which is because in the first time in _CI_OK() we
will create a new ci->cachedsc without copying the name of the old one.