Fix reusing glyph classes with UseMarkFilteringSet
If a class is already defined we will try to use it, but this fails with
mark filtering sets and error is raised. We seem to be returning the
wrong value here. I don’t have the slightest idea what this code does,
but this machos the similar return at the end of this function.