Fix crash in PyFFContour_draw()
The code was trying to “rotate” the points so that it always starts with
on curve point, and `off` here is supposed be the last point from the
loop above it, so it should be `i - start` otherwise we will be writing
past the allocated array while leaving parts of it uninitialized causing
a crash later. This only happens if the contour is starting with off
curve point.